+++ src/Delta/Html/Compound.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace DBlackborough\Quill\Delta\Html;
 
 use DBlackborough\Quill\Options;
+use DBlackborough\Quill\Settings;
 
 /**
  * Compound HTML delta, collects all the attributes for a compound insert and returns the generated HTML
@@ -82,7 +83,9 @@ class Compound extends Delta
                     break;
 
                 default:
-                    $this->element_attributes[$attribute] = $value;
+                    if (in_array($attribute, Settings::ignoredCustomAttributes()) === false) {
+                        $this->element_attributes[$attribute] = $value;
+                    }
                     break;
             }
         }
@@ -123,10 +126,12 @@ class Compound extends Delta
 
         $element_attributes = '';
         foreach ($this->element_attributes as $attribute => $value) {
-            if ($attribute == "color") {
-                $element_attributes .= "style=\"{$attribute}: $value\"";
-            } else {
-                $element_attributes .= "{$attribute}=\"{$value}\" ";
+            if (is_string($attribute) && is_string($value)) {
+                if ($attribute == "color") {
+                    $element_attributes .= "style=\"{$attribute}: $value\"";
+                } else {
+                    $element_attributes .= "{$attribute}=\"{$value}\" ";
+                }
             }
         }